Phase 4: Complete a function to convert integers to Roman Numerals.
If you get user input in the form of Roman Numerals, you will need to convert them into integers to do the calculations. But once the calculations are done, you will want to convert them back into Roman Numerals so that the user can read them.

Create a function called intToRoman() that takes an integer as input and converts it to Roman Numerals, and returns the converted Roman Numerals.